requirements.txt
pillow
pysimplegui
click
requests
pyspng
ninja
imageio-ffmpeg
dlib
numpy
torch